refactor(sens): the sensitivity core is pounce's, pyomo-pounce is a caller
Every numeric behind sens_solution(), sens_solution_report(),
sens_active_set_changes(), sens_covariance() and sens_information()
lived in pyomo-pounce/pyomo_pounce/sens.py and was reachable only with
Pyomo installed and a Pyomo model in hand. A .nl, CLI or CasADi user had
the linear step and the reduced Hessian and nothing else: no fix-relax,
no path mode, no corrector, no activity classification, no covariance,
no identifiability.

None of that code was Pyomo-shaped. Measured before the move, 82 of its
100 functions carried no Pyomo reference in executable code; the
headline five -- sens_covariance (454 lines), sens_solution (312),
_classify_fitted_block (240), sens_information (231),
sens_active_set_changes (96) -- carried zero, and
sens_solution_report's only two were the ComponentMap() it built its
answer in. It already worked in row indices and numpy. It was packaged
Pyomo-first, not written that way.

2400 lines move to pounce/sensitivity/, addressed by row:

    nl   = pounce.read_nl("model.nl")
    sess = solve_for_sensitivity(nl, pins={"p": 4})
    solution(sess, [4], [0.05], mode="fix_relax")

SensSession is the seam -- a solved NL plus its held factor, keyed by
whatever the caller keys it with. pyomo_pounce._Session subclasses it
and supplies three hooks (var_key, user_row_data, new_keymap), so a
Pyomo caller still gets ComponentMaps of component data back and a
bare-NL caller gets .col names. `who=` keeps each wrapper's own function
name in the messages its users see, and a `hints` mapping keeps the
declarations those messages point at spelled the caller's way
(declare_sens_residual() for Pyomo, res_rows= for everyone else), so no
diagnostic names an API the reader is not using.

The Pyomo API is unchanged: same functions, arguments, result types
(Covariance, Information, SolutionReport, ActiveSetChange are
re-exported from the core) and messages. Its 500 tests pass untouched
apart from five imports of internals that moved.

python/tests/test_sensitivity_core.py is the point of the exercise:
twelve tests, no Pyomo, checked against closed forms rather than a
previous run -- the step against dx/dp = 1, df/dp against 6p on the
gh#878 model where a chain-rule-only reading returns 0, the covariance
against the linear least-squares formula, the information matrix against
2*sum(t^2). Two exist because this refactor creates a package boundary
exactly where full-x meets var-x: a fixture with an equal-bounds
variable AHEAD of the parameter makes the spaces genuinely diverge
(_primal_row_map() == [0, 1, None, 2]) and pins the step across it,
since reading one as the other returns a neighbouring variable's answer
-- plausible, wrong, and gh#450 twice already.

Two duplications go with it. _NlBridge is now
pounce.sensitivity.NlBridge, used by both. And _curve_fit.py and the
covariance code each carried their own SVD nullspace basis -- same
formula, same rank tolerance, written twice; both now call
pounce._stats_util.nullspace.

Their covariance routines are deliberately NOT merged: curve_fit reads
inv(H_S) straight off the factor because its parameters are the decision
variables and K = H_S, while pounce.sensitivity recovers a tangent map
because its fitted block sits inside a larger model with equalities.
Same word, different computation; unifying them would replace a correct
special case with a general routine carrying a guard it cannot need.

pyomo-pounce now requires pounce-solver>=0.11.0, the release this lands
in: an older wheel imports but every sensitivity entry point raises
ImportError. That floor also subsumes the pyomo-v2 extra's
pounce-solver>0.9.0, so the extra's comment no longer claims both its
floors are tighter than the base -- the constraint stays, because it
records why the v2 route needs it (gh#553).
